In monte_carlo I added the function transport_neutron which creates a neutron and follows it until it is absorbed, calculating its total distance travelled.
In circle I created a function that calculates the value of pi by randomly placing darts in a square and seeing how many land in a circle inscribed within the square.

In assignment I determined that the clumping of points on the second plot was due to a bad angle parameter.